backend/requirements.txt
fastapi
uvicorn
supabase
pydantic
python-dotenv
python-multipart
passlib[argon2]
